+++ src/main/kotlin/dartzee/utils/ResourceCache.kt
@@ -60,6 +60,7 @@ object ResourceCache
     val URL_ACHIEVEMENT_DARTZEE_UNDER_PRESSURE: URL? = javaClass.getResource("/achievements/underPressure.png")
     val URL_ACHIEVEMENT_DARTZEE_BINGO: URL? = javaClass.getResource("/achievements/bingo.png")
     val URL_ACHIEVEMENT_DARTZEE_HALVED: URL? = javaClass.getResource("/achievements/bust.png")
+    val URL_ACHIEVEMENT_X01_CHUCKLEVISION: URL? = javaClass.getResource("/achievements/chucklevision.png")
 
     val BASE_FONT: Font = Font.createFont(Font.TRUETYPE_FONT, javaClass.getResourceAsStream("/trebuc.ttf"))
